third_party/WebKit/LayoutTests/http/tests/serviceworker/chromium/service-worker-gc-expected.txt

Test that a registered Service Worker with an event handler is not garbage collected prematurely

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS swObservation.wasCollected is false
PASS swObservation.wasCollected is true
PASS successfullyParsed is true

TEST COMPLETE
